+++ cognite/client/data_classes/assets.py
@@ -9,18 +9,18 @@ class Asset(CogniteResource):
     """A representation of a physical asset, for example a factory or a piece of equipment.
 
     Args:
-        external_id (str): The external ID provided by the client. Must be unique within the project.
+        external_id (str): The external ID provided by the client. Must be unique for the resource type.
         name (str): The name of the asset.
-        parent_id (int): A JavaScript-friendly internal ID for the object.
+        parent_id (int): A server-generated ID for the object.
         description (str): The description of the asset.
         metadata (Dict[str, Any]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
         source (str): The source of the asset.
-        id (int): A JavaScript-friendly internal ID for the object.
+        id (int): A server-generated ID for the object.
         created_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
         last_updated_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
-        root_id (int): A JavaScript-friendly internal ID for the object.
+        root_id (int): A server-generated ID for the object.
         aggregates (Dict[str, Any]): Aggregated metrics of the asset
-        parent_external_id (str): The external ID provided by the client. Must be unique within the project.
+        parent_external_id (str): The external ID provided by the client. Must be unique for the resource type.
         cognite_client (CogniteClient): The client to associate with this object.
     """
 
@@ -139,8 +139,8 @@ class AssetUpdate(CogniteUpdate):
     """Changes applied to asset
 
     Args:
-        id (int): A JavaScript-friendly internal ID for the object.
-        external_id (str): The external ID provided by the client. Must be unique within the project.
+        id (int): A server-generated ID for the object.
+        external_id (str): The external ID provided by the client. Must be unique for the resource type.
     """
 
     @property
@@ -282,14 +282,14 @@ class AssetFilter(CogniteFilter):
 
     Args:
         name (str): The name of the asset.
-        parent_ids (List[int]): No description.
-        root_ids (List[Union[Dict[str, Any], Dict[str, Any]]]): No description.
+        parent_ids (List[int]): Return only the direct descendants of the specified assets.
+        root_ids (List[Union[Dict[str, Any], Dict[str, Any]]]): Return all descendants of the specified root assets.
         metadata (Dict[str, Any]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
         source (str): The source of the asset.
         created_time (Dict[str, Any]): Range between two timestamps.
         last_updated_time (Dict[str, Any]): Range between two timestamps.
-        root (bool): Whether the listed assets are root assets, or not.
-        external_id_prefix (str): The external ID provided by the client. Must be unique within the project.
+        root (bool): Whether the filtered assets are root assets, or not. Set to True to only list root assets.
+        external_id_prefix (str): The external ID provided by the client. Must be unique for the resource type.
         cognite_client (CogniteClient): The client to associate with this object.
     """