+++ cognite/client/data_classes/time_series.py
@@ -168,8 +168,8 @@ class TimeSeriesUpdate(CogniteUpdate):
     """Changes will be applied to time series.
 
     Args:
-        id (int): A JavaScript-friendly internal ID for the object.
-        external_id (str): The external ID provided by the client. Must be unique within the project.
+        id (int): A server-generated ID for the object.
+        external_id (str): The external ID provided by the client. Must be unique for the resource type.
     """
 
     @property