+++ tests/tests_unit/test_api/test_assets.py
@@ -117,6 +117,40 @@ class TestAssets:
         res = ASSETS_API.list(name="bla",aggregated_properties=["child_count"])
         assert ["childCount"] == jsgz_load(mock_assets_response.calls[0].request.body)["aggregatedProperties"]
 
+    def test_list_root(self, mock_assets_response):
+        ASSETS_API.list(root_ids=[{"id": 1}, {"externalId": "abc"}], limit=10)
+        calls = mock_assets_response.calls
+        assert 1 == len(calls)
+        assert {"cursor": None, "limit": 10, "filter": {"rootIds": [{"id": 1}, {"externalId": "abc"}]}} == jsgz_load(
+            calls[0].request.body
+        )
+
+    def test_call_root(self, mock_assets_response):
+        list(ASSETS_API.__call__(root_ids=[{"id": 1}, {"externalId": "abc"}], limit=10))
+        calls = mock_assets_response.calls
+        assert 1 == len(calls)
+        assert {"cursor": None, "limit": 10, "filter": {"rootIds": [{"id": 1}, {"externalId": "abc"}]}} == jsgz_load(
+            calls[0].request.body
+        )
+
+    def test_list_root_ids_list(self, mock_assets_response):
+        ASSETS_API.list(root_ids=[1, 2], limit=10)
+        calls = mock_assets_response.calls
+        assert 1 == len(calls)
+        assert {"cursor": None, "limit": 10, "filter": {"rootIds": [{"id": 1}, {"id": 2}]}} == jsgz_load(
+            calls[0].request.body
+        )
+
+    def test_list_root_extids_list(self, mock_assets_response):
+        ASSETS_API.list(root_external_ids=["1", "2"], limit=10)
+        calls = mock_assets_response.calls
+        assert 1 == len(calls)
+        assert {
+            "cursor": None,
+            "limit": 10,
+            "filter": {"rootIds": [{"externalId": "1"}, {"externalId": "2"}]},
+        } == jsgz_load(calls[0].request.body)
+
     def test_create_single(self, mock_assets_response):
         res = ASSETS_API.create(Asset(external_id="1", name="blabla"))
         assert isinstance(res, Asset)