+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Changes are grouped as follows
 - `Asset` and `AssetList` classes now have a `sequences` function which retrieves related sequences.
 
 ### Changed
+- `list` and `__call__` methods for assets now support list parameters for `root_ids`, `root_external_ids`.
+
+### Changed
 - Sequences API no longer supports column ids, all relevant functions have been changed to only use external ids.
 
 ### Fixed